Moved classes in ObjectCreatorDrawer file to proper files. Simplified RobotAgent initialization. Moved ObjectPositionSettings to MonoBehaviour. Simplified code in RandomInitOnMesh cless.
