--- Assets/Scripts/Objects/IRandomInit.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace Objects
-{
-    interface IRandomInit
-    {
-        void PutAll();
-        void PutTarget(GameObject target);
-    }
-}
