--- Assets/Scripts/Objects/IRandomInit.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 22f3685842fd5ac448c8b7c0e39e6ed7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
