+++ Assets/Scripts/Objects/ObjectCreator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3038ed21f12dcc6c6b7771790963dcbc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
