--- Assets/Scripts/Objects/ObjectCreatorDrawer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d891df1fd9dc47c45846223b89a07a80
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
