+++ Assets/Scripts/Objects/ObjectPositionSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6913e23d5dab23a4cbef4154cd843a5a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
