+++ Assets/Scripts/Objects/RandomInitOnMesh.cs
@@ -4,48 +4,30 @@ using UnityEngine;
 
 namespace Objects
 {
-    public class RandomInitOnMesh : MonoBehaviour, IRandomInit
+    public class RandomInitOnMesh : MonoBehaviour
     {
-        private Renderer mesh;
-        private RaycastHit hit;
-        private Vector3 position;
+        public Renderer mesh;
         private int layerMask = 1 << 10;
         [SerializeField]
         private List<GameObject> objectsToPut = new List<GameObject>();
 
-        Vector3 GetRandomPosition(GameObject target)
+        Vector3 GetPosition()
         {
-            float randomX = Utils.GetRandom(mesh.bounds.min.x, mesh.bounds.max.x);
-            float randomZ = Utils.GetRandom(mesh.bounds.min.z, mesh.bounds.max.z);
-            float y_pos = 0;
+            Vector3 position = new Vector3();
+            position.x=Utils.GetRandom(mesh.bounds.min.x, mesh.bounds.max.x);
+            position.z=Utils.GetRandom(mesh.bounds.min.z, mesh.bounds.max.z);
+            position.y=mesh.bounds.max.y;
+            RaycastHit hit;
             if (Physics.Raycast(position, -Vector3.up, out hit, Mathf.Infinity, layerMask))
             {
-                y_pos = -hit.distance + 2f;
+                position.y = -hit.distance + 2f;
             }
-            return new Vector3(randomX, target.transform.position.y, randomZ);
-        }
-
-        bool IsCorrectPosition()
-        {
-            if (!Physics.Raycast(position, -Vector3.up, out hit, Mathf.Infinity, layerMask))
-                return false;
-            return true;
-        }
-
-        Vector3 GetPosition(GameObject target)
-        {
-            do
-            {
-                position = GetRandomPosition(target);
-            } while (!IsCorrectPosition());
-
             return position;
         }
 
         public void PutTarget(GameObject target)
         {
-            position = GetPosition(target);
-            target.transform.position = position;
+            target.transform.position = GetPosition();
         }
 
         public void PutAll()
